[기능요청][FCM] FCM 알림 서비스 move: NotificationType 클래스 위치 변경 #104
